requirements.txt
alabaster==1.0.0
anyio==4.12.1
argcomplete==3.6.2
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asttokens==3.0.1
async-lru==2.2.0
attrs==25.4.0
autocommand==2.2.2
babel==2.17.0
beautifulsoup4==4.14.3
bleach==6.3.0
Brlapi==0.8.7
btrfsutil==6.19
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
comm==0.2.3
configobj==5.0.9
cryptography==46.0.5
dasbus==1.7
dbus-python==1.4.0
decorator==5.2.1
defusedxml==0.7.1
distlib==0.4.0
docutils==0.22.4
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.24.0
GDAL==3.12.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
imagesize==1.4.1
importlib_metadata==8.7.1
iniconfig==2.2.0
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==4.0.0
jedi==0.19.2
Jinja2==3.1.6
json5==0.13.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyterlab==4.5.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
lensfun==0.3.4
libfdt==1.7.2
libvirt-python==12.1.0
louis==3.36.0
lxml==6.0.2
Mako==1.3.10.dev0
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
meson==1.10.2
mistune==3.2.0
more-itertools==10.8.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest_asyncio==1.5.9
netsnmp-python==1.0a1
nftables==0.1
notebook_shim==0.2.4
numpy==2.4.3
OWSLib==0.34.1
packaging==26.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
perf==0.1
pexpect==4.9.0
pillow==12.1.1
pkg_resources==81.0.0
platformdirs==4.9.4
pluggy==1.6.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
psutil==7.2.2
psutils==3.3.14
psycopg2==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==2.0.2
pwquality==1.4.5
pycairo==1.29.0
pycparser==2.23
Pygments==2.19.2
PyGObject==3.54.5
pyliblo3==0.16.3
pyparsing==3.3.2
pypdf==6.8.0
PyQt5==5.15.11
PyQt5_sip==12.18.0
pytest==8.4.2
python-dateutil==2.9.0
python-json-logger==4.0.0
python-linux-procfs==0.7.4
pytz==2025.2
pyudev==0.24.4
pyxdg==0.28
PyYAML==6.0.3
pyzmq==27.1.0
QScintilla==2.14.1
rdflib==7.6.0
referencing==0.37.0
requests==2.32.5
rich==14.3.3
roman-numerals-py==3.1.0
rpds-py==0.30.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.15
Send2Trash==1.8.3
setproctitle==1.3.7
setuptools==82.0.1
six==1.17.0
sniffio==1.3.1
snowballstemmer==3.0.0.1
soupsieve==2.8.3
Sphinx==9.1.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack_data==0.6.4.dev0+g53501eb8b.d20251216
systemd-python==235
tinycss2==1.5.1
tornado==6.5.3
tqdm==4.67.3
traitlets==5.14.3
typing_extensions==4.15.0
urllib3==2.6.3
uv==0.10.10
validate==5.0.9
wcwidth==0.6.0
webencodings==0.5.1
websocket-client==1.9.0
wheel==0.46.3
zipp==3.21.0
